Refactor AddMetricPopup to screen dialog
